A MultiHttpServer and keepalive was added
A MultiHttpServer that can handle up to 10 clients was introduced as
well as a function to check for a keepalive request from the client. In
this case if the client adds 'yes' to the end of a request, the server
will wait 5 seconds before closing the connection. If the client does
not add a request, the default case is for the server to close the
connection immediately after the output data is sent.